fba_backend/requirements.txt
apispec==5.2.2
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==2.2.5
Flask-Cors==5.0.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
flask-smorest==0.36.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mongoengine==0.29.1
packaging==24.2
pymongo==4.11
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0
webargs==8.6.0
Werkzeug==2.3.8
WTForms==3.2.1
gunicorn==23.0.0
